+++ repository/book_repository.go
@@ -7,8 +7,8 @@ import (
 )
 
 type BookRepository interface {
-	InsertBook(ctx context.Context, medicalRecordNumber int32, time int) (int32, error)
-	FindBookByMedicalRecordNumber(ctx context.Context, medicalRecordNumber int32) ([]entity.Book, error)
-	FindBookById(ctx context.Context, bookId int32, medicalRecordNumber int32) error
-	UpdateBook(ctx context.Context, book entity.Book, medicalRecordNumber int32, bookid int32) (int32, error)
+	InsertBook(ctx context.Context, medicalRecordNumber int64, time int) (int64, error)
+	FindBookByMedicalRecordNumber(ctx context.Context, medicalRecordNumber int64) ([]entity.Book, error)
+	FindBookById(ctx context.Context, bookId int32, medicalRecordNumber int64) error
+	UpdateBook(ctx context.Context, book entity.Book, medicalRecordNumber int64, bookid int32) (int32, error)
 }
